Separated the tests for setTypeMap and getTypeMap from the inverted map tests.
Created a new test suite for user-defined data types, since there will be quite a few tests.
