+++ pgjdbc/src/test/java/org/postgresql/test/jdbc2/Jdbc2TestSuite.java
@@ -133,9 +133,7 @@ import org.junit.runners.Suite;
         CopyLargeFileTest.class,
         ServerErrorTest.class,
         UpsertTest.class,
-        OuterJoinSyntaxTest.class,
-
-        TypeMapTest.class
+        OuterJoinSyntaxTest.class
 })
 public class Jdbc2TestSuite {
 }
